perf(app): organize handlers based on vertical slice
