+++ quetzalcoatl-auth/Application/Application.csproj
@@ -5,6 +5,7 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
   <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="12.0.1" />
     <PackageReference Include="FastEndpoints" Version="5.8.0" />
     <PackageReference Include="FastEndpoints.Security" Version="5.8.0" />
   </ItemGroup>
