+++ quetzalcoatl-auth/Application/Features/Jwt/GenerateJwtToken/Command.cs
@@ -1,0 +1,6 @@
+namespace Application.Features.Jwt.GenerateJwtToken;
+
+public class GenerateJwtTokenCommand : ICommand<string>
+{
+    public ApplicationUser User { get; set; } = default!;
+}
