+++ quetzalcoatl-auth/Application/Features/Users/CreateUser/Command.cs
@@ -1,0 +1,8 @@
+namespace Application.Features.Users.CreateUser;
+
+public class CreateUserCommand : ICommand<ApplicationUser>
+{
+    public string UserName { get; set; } = default!;
+    public string Email { get; set; } = default!;
+    public string Password { get; set; } = default!;
+}
