+++ quetzalcoatl-auth/Application/Features/Users/CreateUser/Mappers.cs
@@ -1,0 +1,11 @@
+namespace Application.Features.Users.CreateUser;
+
+public class CreateUserCommandToApplicationUserProfile : Profile
+{
+    public CreateUserCommandToApplicationUserProfile()
+    {
+        CreateMap<CreateUserCommand, ApplicationUser>()
+            .ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.UserName))
+            .ForMember(dest => dest.Email, opt => opt.MapFrom(src => src.Email));
+    }
+}
