+++ quetzalcoatl-auth/Application/Features/Users/ValidateUserCredentials/Command.cs
@@ -1,0 +1,7 @@
+namespace Application.Features.Users.ValidateUserCredentials;
+
+public class ValidateUserCredentialsCommand : ICommand<ApplicationUser>
+{
+    public string Email { get; set; } = default!;
+    public string Password { get; set; } = default!;
+}
