+++ quetzalcoatl-auth/Application/IApplicationMarker.cs
@@ -1,0 +1,3 @@
+namespace Application;
+
+public interface IApplicationMarker { }
