+++ quetzalcoatl-auth/Application/Usings.cs
@@ -4,3 +4,5 @@ global using Microsoft.AspNetCore.Identity;
 global using FastEndpoints.Security;
 global using System.IdentityModel.Tokens.Jwt;
 global using System.Security.Claims;
+global using AutoMapper;
+global using IMapper = AutoMapper.IMapper;
