+++ quetzalcoatl-auth/Bootstrapper/Usings.cs
@@ -3,6 +3,6 @@ global using Infrastructure;
 global using Microsoft.EntityFrameworkCore;
 global using Microsoft.AspNetCore.Identity;
 global using FastEndpoints.Security;
-global using Api.Auth.Register;
 global using Api;
 global using Domain.Entities;
+global using Application;
